[Issue-3967] Update UX for Welcome screen
